Converted to NodeJS using ES6 modules.
