+++ index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>Vector Class</title>
+</head>
+<body>
+  <h1>Testing...</h1>
+  <!-- This will run on Node, but not in browser -->
+    <script type="text/javascript" src="./test.mjs"></script>
+  <!-- End test -->
+</body>
+</html>
